+++ Adapters/BidonAdapterMoloco/BidonAdapterMoloco.h
@@ -1,0 +1,14 @@
+//
+//  BidonAdapterMoloco.h
+//  BidonAdapterMoloco
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for BidonAdapterMoloco.
+FOUNDATION_EXPORT double BidonAdapterMolocoVersionNumber;
+
+//! Project version string for BidonAdapterMoloco.
+FOUNDATION_EXPORT const unsigned char BidonAdapterMolocoVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <BidonAdapterMoloco/PublicHeader.h>