+++ Adapters/BidonAdapterMoloco/Models/MolocoAdUnitExtras.swift
@@ -1,0 +1,15 @@
+//
+//  MolocoAdUnitExtras.swift
+//  BidonAdapterMoloco
+//
+//  Created by Andrei Rudyk on 20/08/2025.
+//
+
+import Foundation
+
+
+struct MolocoAdUnitExtras: Codable {
+    var adUnitId: String
+}
+
+struct MolocoBiddingTokenExtras: Codable { }