+++ Adapters/BidonAdapterMoloco/Models/MolocoBiddingResponse.swift
@@ -1,0 +1,13 @@
+//
+//  MolocoBiddingResponse.swift
+//  BidonAdapterMoloco
+//
+//  Created by Andrei Rudyk on 20/08/2025.
+//
+
+import Foundation
+
+
+struct MolocoBiddingResponse: Codable {
+    var payload: String
+}