+++ Adapters/BidonAdapterMoloco/Models/MolocoParameters.swift
@@ -1,0 +1,12 @@
+//
+//  MolocoParameters.swift
+//  BidonAdapterMoloco
+//
+//  Created by Andrei Rudyk on 19/08/2025.
+//
+
+import Foundation
+
+public struct MolocoParameters: Codable {
+    var appKey: String
+}