+++ Bidon/SDK/Fullscreen/InterstitialImpressionController.swift
@@ -35,7 +35,7 @@ final class InterstitialImpressionController: NSObject, FullscreenImpressionCont
     func show(from context: UIViewController) {
         provider.show(opaque: impression.ad, from: context)
     }
-    
+
     func notifyWin() {
         guard bid.adUnit.bidType == .direct else {
             Logger.info("[Win/Loss] Do not notify \(bid.adUnit.demandId) win because it's bid type is RTB")
@@ -44,7 +44,7 @@ final class InterstitialImpressionController: NSObject, FullscreenImpressionCont
         Logger.info("[Win/Loss] Notify \(bid.adUnit.demandId) win")
         provider.notify(opaque: bid.ad, event: .win)
     }
-    
+
     func notifyLose(winner demandId: String, eCPM: Price) {
         guard bid.adUnit.bidType == .direct else {
             Logger.info("[Win/Loss] Do not notify \(demandId) lose because it's bid type is RTB")