+++ Bidon/Shared/Constants.swift
@@ -9,7 +9,7 @@ import Foundation
 
 
 struct Constants {
-    static let sdkVersion: String = "0.10.0"
+    static let sdkVersion: String = "0.11.0"
 
     static let zeroUUID: String = "00000000-0000-0000-0000-000000000000"
 
@@ -39,7 +39,8 @@ struct Constants {
             "BidonAdapterMyTarget.MyTargetDemandSourceAdapter",
             "BidonAdapterIronSource.IronSourceDemandSourceAdapter",
             "BidonAdapterYandex.YandexDemandSourceAdapter",
-            "BidonAdapterChartboost.ChartboostDemandSourceAdapter"
+            "BidonAdapterChartboost.ChartboostDemandSourceAdapter",
+            "BidonAdapterMoloco.MolocoDemandSourceAdapter"
         ]
     }